A lot of added features
The tutorial walked me through adding quite a few features. The page now responds with updated ingredient amount data based on what the user adjusts the servings to. The list of recipes on the side highlights the current one that has been selected and the information for the recipe populates to the middle of the page.