Finished
The shopping list ingredients and amounts appears on the right side of the page when a recipe is added to the cart. The amount of ingredients can be changed. User can like recipes, which stores them in the liked recipes heart in the top right corner via local storage. There are features I can add/improve on to make this tutorial my own, but I am done with it for now. I was unable to load the node_modules folder because it was way too much data.